requirements.txt
annotated-types==0.7.0
colorama==0.4.6
colorlog==6.8.0
dnspython==2.7.0
greenlet==3.2.2
kafka-python==2.0.2
numpy==2.2.6
pandas==2.2.3
psycopg2-binary
pydantic==2.11.5
pydantic-settings==2.1.0
pydantic_core==2.33.2
pymongo==4.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
redis==5.0.1
six==1.17.0
SQLAlchemy==2.0.23
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
